multimodal-search-poc-backend/requirements.txt
# Base dependencies
numpy==1.26.1
scipy==1.10.1
Pillow>=10.1.0
requests==2.31.0
tqdm==4.66.1
aiohttp==3.9.1
python-dotenv==1.0.0
python-multipart==0.0.6
python-jose==3.3.0
scikit-learn>=1.0.2

# PyTorch ecosystem (already installed)
# torch==1.13.1
# torchvision==0.14.1

# ML tools
faiss-cpu==1.7.4
transformers==4.34.1
sentence-transformers==2.2.2

# FastAPI
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.1.0

# Audio processing
librosa==0.10.1
soundfile==0.12.1

# Text processing
python-Levenshtein==0.23.0
